Fixed all changes
-bug tested the auto and fixed direction
