+++ src/main/java/frc/robot/commands/Autos/Autos.java
@@ -27,7 +27,7 @@ public final class Autos {
                 new DoNothing("Doing Nothing"));
 
         commandChooser.addOption("Drive Forward",
-                new DriveForwardOG(3, RobotContainer.drivebase));
+                new DriveForward(3, RobotContainer.drivebase));
 
                 
         commandChooser.addOption("Reef Top",
