+++ src/main/java/frc/robot/commands/Autos/DriveForward.java
@@ -4,14 +4,50 @@
 
 package frc.robot.commands.Autos;
 
+import edu.wpi.first.math.kinematics.ChassisSpeeds;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+import edu.wpi.first.wpilibj2.command.WaitCommand;
+import frc.robot.RobotContainer;
+import frc.robot.subsystems.swervedrive.SwerveSubsystem;
 
-// NOTE:  Consider using this command inline, rather than writing a subclass.  For more
-// information, see:
-// https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class DriveForward extends SequentialCommandGroup {
-  /** Creates a new DriveForward2. */
-  public DriveForward() {
-    addCommands();
+public class DriveForward extends WaitCommand {
+  int wait;
+  final int stop = 0;
+  SwerveSubsystem swerveSubsystem;
+
+  public DriveForward(double seconds, SwerveSubsystem ss) {
+    super(seconds);
+    swerveSubsystem = ss;
+    addRequirements(ss);
+
+  }
+
+  @Override
+  public void initialize() {
+    super.initialize();
+    wait = 0;
+    swerveSubsystem.zeroGyroWithAlliance();
+    System.out.println("DriveForward initialize");
+  }
+
+  // Called every time the scheduler runs while the command is scheduled.
+  @Override
+  public void execute() {
+    var direction = 1.0;
+
+    if (swerveSubsystem.isRedAlliance()) {
+
+      direction = -1.0;
+    }
+
+    swerveSubsystem.driveFieldOriented(new ChassisSpeeds(-0.5 * direction, 0, 0));
+    wait += 1;
+  }
+
+  // Called once the command ends or is interrupted.
+  @Override
+  public void end(boolean interrupted) {
+    swerveSubsystem.driveFieldOriented(new ChassisSpeeds());
+    System.out.println("Ending Drive Forward");
   }
 }
